+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Endpoints to get practice codes and names for EQIP and CSP. [#263](https://github.com/policy-design-lab/pdl-api/issues/263)
 - Total payment grouped by practice codes to state distribution endpoint for EQIP and CSP, with query parameter support. [#275](https://github.com/policy-design-lab/pdl-api/issues/275)
+- Add house outlay to Title II endpoint. [#280](https://github.com/policy-design-lab/pdl-api/issues/280)
 
 ### Fixed
 - SNAP average monthly participation and percentage calculation across years. [#278](https://github.com/policy-design-lab/pdl-api/issues/278)